ログイン機能の実装

+++ JRP/src/main/java/com/app/controller/AD0001Controller.java
@@ -27,7 +27,9 @@ import com.app.dto.AD0001.AD0001DownloadDto;
 import com.app.dto.AD0001.AD0001Dto;
 import com.app.dto.AD0001.AD0001UpdateDto;
 import com.app.logic.AD0001Logic;
+import com.app.logic.ST0001Logic;
 import com.google.api.core.ApiFuture;
+import com.google.cloud.firestore.DocumentReference;
 import com.google.cloud.firestore.Firestore;
 import com.google.cloud.firestore.QueryDocumentSnapshot;
 import com.google.cloud.firestore.QuerySnapshot;
@@ -41,6 +43,9 @@ public class AD0001Controller {
 	
 	@Autowired
 	private AD0001Logic logic;
+	
+    @Autowired
+    private ST0001Logic st0001Logic;
 	
 	@Value("${gcs.bucket.name}")
     private String bucketName;
@@ -79,23 +84,48 @@ public class AD0001Controller {
 	
 	@PostMapping(ActionName.INSERT)
 	public ResponseDto create(@RequestBody AD0001CreateDto dto) {
-			try {
-				firestore.collection("users").add(dto).get();
-			} catch (InterruptedException e) {
-				// TODO 自動生成された catch ブロック
-				e.printStackTrace();
-			} catch (ExecutionException e) {
-				// TODO 自動生成された catch ブロック
-				e.printStackTrace();
-			}
+
+		if (dto.getName() == null || dto.getName().equals("さん")) {
+			ResponseDto response = new ResponseDto();
+            response.setSuccess(false);
+            response.setMessage("ユーザー名を入力してください。");
+            return response;
+        }
+
+		try {
+			DocumentReference documentReference = firestore.collection("users").add(dto).get();
+
+			st0001Logic.createMotivation(documentReference.getId());
+		} catch (InterruptedException e) {
+			// TODO 自動生成された catch ブロック
+			e.printStackTrace();
+		} catch (ExecutionException e) {
+			// TODO 自動生成された catch ブロック
+			e.printStackTrace();
+		}
 
 		ResponseDto response = new ResponseDto();
 		response.setSuccess(true);
+		response.setMessage("登録しました。");
 		return response;
 	}
 	
 	@PostMapping(ActionName.UPDATE)
 	public ResponseDto update(@RequestBody AD0001UpdateDto dto) {
+		if (dto.getId() == null || dto.getId().isEmpty()) {
+			ResponseDto response = new ResponseDto();
+            response.setSuccess(false);
+            response.setMessage("ユーザーを選択してください。");
+            return response;
+        }
+		
+		if (dto.getName() == null || dto.getName().equals("さん")) {
+			ResponseDto response = new ResponseDto();
+            response.setSuccess(false);
+            response.setMessage("ユーザー名を入力してください。");
+            return response;
+        }
+		
 		try {
 			firestore.collection("users")
 				.document(dto.getId())
@@ -113,15 +143,25 @@ public class AD0001Controller {
 
 		ResponseDto response = new ResponseDto();
 		response.setSuccess(true);
+		response.setMessage("更新しました。");
 		return response;
 	}
 	
 	@PostMapping(ActionName.DELETE)
 	public ResponseDto delete(@RequestBody AD0001DeleteDto dto) {
+		if (dto.getId() == null || dto.getId().isEmpty()) {
+			ResponseDto response = new ResponseDto();
+            response.setSuccess(false);
+            response.setMessage("ユーザーを選択してください。");
+            return response;
+        }
+
         try {
             firestore.collection("users")
                 .document(dto.getId())
                 .delete().get();
+            
+           st0001Logic.deleteMotivation(dto.getId());
         } catch (InterruptedException e) {
             // TODO 自動生成された catch ブロック
             e.printStackTrace();
@@ -132,6 +172,7 @@ public class AD0001Controller {
 
         ResponseDto response = new ResponseDto();
         response.setSuccess(true);
+        response.setMessage("削除しました。");
         return response;
 
 	}
